requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
apify_client==1.6.2
apify_shared==1.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.6.2
boto3==1.34.19
botocore==1.34.19
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
clickhouse-connect==0.5.25
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
duckdb==0.9.2
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.1.0
fastapi==0.109.2
filelock==3.13.1
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.27.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langchainplus-sdk==0.0.16
langsmith==0.0.83
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.20.2
marshmallow-enum==1.5.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.9.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pandas==2.0.2
pipenv==2023.11.15
platformdirs==4.1.0
pluggy==1.0.0
posthog==3.4.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pypandoc==1.12
pypdf==4.0.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytest==7.3.2
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.3.4
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.21.0
